Bug Fixes
Fixed: Normal players cant open sandbot inventory if raid feature enabled
Fixed: Enemy faction players can bypass remove sandbot gui

+++ src/main/java/me/magas8/GUIS/SandBotGui.java
@@ -123,7 +123,7 @@ public class SandBotGui extends MenuManager implements InventoryProvider {
             Player invplayer = (Player) e.getWhoClicked();
             double balanceplayer = LunarSandBot.econ.getBalance(player);
             double balancebot = bot.getBalance();
-            double howmanytotake = (double) plugin.getConfig().getInt("money-to-add");
+            double howmanytotake = (double) plugin.getConfig().getDouble("money-to-add");
             if(LunarSandBot.botGuis.containsKey(bot)) {
                 if (e.isLeftClick()) {
